% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouped_glance.R
\name{grouped_glance}
\alias{grouped_glance}
\title{Model summary from grouped analysis of any function that has \code{data}
argument in its function call.}
\usage{
grouped_glance(data, grouping.vars, ..f, ...)
}
\arguments{
\item{data}{Dataframe from which variables are to be taken.}

\item{grouping.vars}{List of grouping variables.}

\item{..f}{A function, or function name as a string.}

\item{...}{Arguments to function.

These dots support \link{tidy-dots} features.}
}
\description{
This is the most general form of a \code{grouped_} function where any
function can be used to run \code{grouped_} analysis.
}
\examples{
groupedstats::grouped_glance(
  data = ggplot2::diamonds,
  grouping.vars = c(cut, color),
  formula = price ~ carat * clarity,
  ..f = stats::lm
)
}
