% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouped_lm.R
\name{grouped_lm}
\alias{grouped_lm}
\title{Function to run linear model (lm) on across multiple grouping
variables.}
\usage{
grouped_lm(data, grouping.vars, formula, output = "tidy")
}
\arguments{
\item{data}{Dataframe from which variables are to be taken.}

\item{grouping.vars}{List of grouping variables.}

\item{formula}{an object of class \code{"\link{formula}"} (or one that
    can be coerced to that class): a symbolic description of the
    model to be fitted.  The details of model specification are given
    under \sQuote{Details}.}

\item{output}{A character describing what output is exepcted. Two possible
options: \code{"tidy"} (default), which will return the results, or \code{"glance"},
which will return model summaries.}
}
\value{
A tibble dataframe with tidy results from linear model.
}
\description{
Function to run linear model (lm) on across multiple grouping
variables.
}
\examples{

grouped_lm(
  data = mtcars,
  grouping.vars = cyl,
  formula = mpg ~ am*wt,
  output = "tidy"
)

grouped_lm(
  data = ggplot2::diamonds,
  grouping.vars = c(cut, color),
  formula = price ~ carat,
  output = "glance"
)

}
\seealso{
grouped_slr
}
\author{
Indrajeet Patil
}
