% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouped_summary.R
\name{grouped_summary}
\alias{grouped_summary}
\title{Function to get descriptive statistics for multiple variables for all grouping variable levels}
\arguments{
\item{data}{Dataframe from which variables need to be taken.}

\item{grouping.vars}{A list of grouping variables.}

\item{measures}{List variables for which summary needs to computed (only \emph{numeric} variables should be entered).}
}
\value{
Dataframe with descriptive statistics for numeric variables (n, mean, sd, median, min, max)
}
\description{
Function to get descriptive statistics for multiple variables for all grouping variable levels
}
\examples{

# if you have multiple variable for each argument
grouped_summary(data = mtcars, grouping.vars = c(am, cyl), measures = c(wt, mpg))
# if you have just one variable per argument
grouped_summary(data = mtcars, grouping.vars = am, measures = wt)

}
\author{
Indrajeet Patil
}
