% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupedstats-package.R
\docType{package}
\name{groupedstats-package}
\alias{groupedstats}
\alias{groupedstats-package}
\title{groupedstats: Statistical Tests for Multiple Variables Across Multiple Grouping Levels}
\description{
Collection of functions to run statistical tests for multiple variable across multiple groups.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/IndrajeetPatil/groupedstats/issues}
}

}
\author{
\strong{Maintainer}: Indrajeet Patil \email{patilindrajeet.science@gmail.com} (0000-0003-1995-6531)

}
\keyword{internal}
