% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupedstats-package.R
\docType{package}
\name{groupedstats}
\alias{groupedstats}
\alias{groupedstats-package}
\title{\code{groupedstats}}
\description{
Collection of functions to run statistical operations on multiple variables
across multiple grouping variables in a dataframe.
}
\details{
For more documentation, see
\href{https://github.com/IndrajeetPatil/groupedstats}{README} on GitHub.
Vignette preparation in progress.
}
