% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{grouped_tidy}
\alias{grouped_glance}
\alias{grouped_augment}
\alias{\%>\%}
\alias{\%<>\%}
\alias{\%$\%}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{broomExtra}{\code{\link[broomExtra]{grouped_tidy}}, \code{\link[broomExtra]{grouped_glance}}, \code{\link[broomExtra]{grouped_augment}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}, \code{\link[magrittr]{\%<>\%}}, \code{\link[magrittr]{\%$\%}}}
}}

