% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouped_summary.R
\name{grouped_summary}
\alias{grouped_summary}
\title{Function to get descriptive statistics for multiple variables for all
grouping variable levels}
\arguments{
\item{data}{Dataframe from which variables need to be taken.}

\item{grouping.vars}{A list of grouping variables.}

\item{measures}{List variables for which summary needs to computed. If not
spcified, all variables of type specified in the argument \code{measures.type}
will be used to calculate summaries. \strong{Don't} explicitly set
\code{measures.type = NULL} in function call, which will produce an error
because the function will try to find a column in a dataframe named "NULL".}

\item{measures.type}{A character indicating whether summary for \emph{numeric}
("numeric") or \emph{factor/character} ("factor") variables is expected
(Default: \code{measures.type = "numeric"}). This function can't be used for
both numeric \strong{and} variables simultaneously.}

\item{topcount.long}{If \code{measures.type = factor}, you can get the top counts
in long format for plotting purposes. (Default: \code{topcount.long = FALSE}).}
}
\value{
Dataframe with descriptive statistics for numeric variables (n, mean,
sd, median, min, max)
}
\description{
Function to get descriptive statistics for multiple variables for all
grouping variable levels
}
\examples{

# another possibility
groupedstats::grouped_summary(
data = datasets::iris,
grouping.vars = Species,
measures = Sepal.Length:Petal.Width,
measures.type = "numeric"
)

# if you have just one variable per argument, you can also not use `c()`
groupedstats::grouped_summary(
data = datasets::ToothGrowth,
grouping.vars = supp,
measures = len,
measures.type = "numeric"
)

}
\author{
Indrajeet Patil
}
