% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupedstats-package.R
\docType{package}
\name{groupedstats-package}
\alias{groupedstats-package}
\title{\code{groupedstats}}
\description{
Collection of functions to run statistical operations on multiple variables
across multiple grouping variables in a dataframe. This package is retired.
}
\details{
For more, see
\href{https://github.com/IndrajeetPatil/groupedstats}{README} on GitHub.
}
